requirements.txt
pandas==1.5.3
matplotlib==3.7.1
requests==2.31.0
streamlit==1.33.0
psycopg2-binary==2.9.9
